agent/fishermans-wharf-agent-v2/requirements.txt
google-adk==1.25.1
google-cloud-aiplatform==1.138.0
google-genai==1.64.0
pydantic==2.12.5
